Fixed a goofy mistake and now sends the right data to server and creates transactions

+++ java_backend/src/main/java/com/tutuzor/javareact/models/Transaction.java
@@ -13,5 +13,5 @@ import javax.persistence.Id;
 public class Transaction {
     @Id @GeneratedValue private Long id;
     @NonNull private Long accountId;
-    @NonNull private Double amount;
+    @NonNull private Long amount;
 }
